requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
array_record==0.8.3
astunparse==1.6.3
attrs==25.4.0
av==13.1.0
benchmarl @ file:///home/pemb7543/DeC_MACTP/BenchMARL
bokeh==3.8.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chex==0.1.91
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
dm-clrs==2.0.3
dm-haiku==0.0.16
dm-tree==0.1.9
docstring_parser==0.17.0
einops==0.8.1
etils==1.13.0
Farama-Notifications==0.0.4
filelock==3.20.2
flatbuffers==25.12.19
flax==0.12.2
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
gnarl @ file:///home/pemb7543/DeC_MACTP/GNARL-MACTP
google-pasta==0.2.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.2.3
h5py==3.15.1
humanize==4.15.0
hydra-core==1.3.2
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
immutabledict==4.2.2
importlib_metadata==8.7.1
importlib_resources==6.5.2
jax==0.8.2
jaxlib==0.8.2
Jinja2==3.1.6
jmp==0.0.4
keras==3.13.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_collections==1.1.0
ml_dtypes==0.5.4
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
namex==0.1.0
narwhals==2.15.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opt_einsum==3.4.0
optax==0.2.6
optree==0.18.0
orbax-checkpoint==0.11.31
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.5.1
plotly==6.5.1
proglog==0.1.12
promise==2.3
propcache==0.4.1
protobuf==5.29.5
psutil==7.2.1
PuLP==3.3.0
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyglet==1.5.27
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyvers==0.1.0
PyYAML==6.0.3
rdkit==2025.9.3
requests==2.32.5
rich==14.2.0
sb3_contrib==2.7.1
scipy==1.16.3
sentry-sdk==2.49.0
simple-parsing==0.1.7
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
soundfile==0.13.1
stable_baselines3==2.7.1
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensordict==0.10.0
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.17.0
tensorstore==0.1.80
termcolor==3.3.0
tfds-nightly==4.9.8.dev202504060044
toml==0.10.2
toolz==1.1.0
torch==2.6.0
torch-geometric==2.6.1
torchrl==0.10.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
treescope==0.1.10
triton==3.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vmas==1.2.12
wandb==0.23.1
Werkzeug==3.1.5
wrapt==2.0.1
xyzservices==2025.11.0
yarl==1.22.0
zipp==3.23.0
